% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_helpers.R
\name{all_after}
\alias{all_after}
\alias{all_before}
\title{stringr regex helpers}
\usage{
all_after(character)

all_before(character)
}
\arguments{
\item{character}{A character string.}
}
\value{
A vector.
}
\description{
Identifies sub-string occuring before or after character.
Note: these are technically all_after_last() and all_before_first() occurrences.
}
\section{Functions}{
\itemize{
\item \code{all_after()}: Extract all characters after match.

\item \code{all_before()}: Extract all characters before match.

}}
\examples{
stringr::str_extract('string_test', all_after("_"))
stringr::str_remove('string_test',  all_after("_"))

stringr::str_extract('string/test', all_after("/"))
stringr::str_extract('string/test', all_before("/"))
stringr::str_extract('string_test', all_before("_"))
stringr::str_remove('string_test',  all_before("_"))
stringr::str_extract('string_test_two', all_after("_"))
stringr::str_remove('string_test_two',  all_after("_"))
stringr::str_extract('string_test_two', all_before("_"))
stringr::str_remove('string_test_two',  all_before("_"))
}
