% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_bottom.R
\name{bottom_legend}
\alias{bottom_legend}
\title{Bottom legend}
\usage{
bottom_legend()
}
\description{
Moves ggplot2 legend to bottom of plotting space.
}
\details{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
