% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_tidyverse_conflicts.R
\name{fix_tidyverse_conflicts}
\alias{fix_tidyverse_conflicts}
\title{Fix tidyverse conflicts}
\usage{
fix_tidyverse_conflicts()
}
\description{
Find functions with name conflicts in the global environment and reassign all, prioritizing tidyverse functions.
}
