% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_origin.R
\name{geom_origin}
\alias{geom_origin}
\title{Geom origin. Draw x-y crosshair through the origin.}
\usage{
geom_origin(lty = 2, xorig = 0, yorig = 0)
}
\arguments{
\item{lty}{line type (e.g. 2 specifies a dashed line)}

\item{xorig}{intercept for x axis}

\item{yorig}{intercept for y axis}
}
\description{
Geom origin. Draw x-y crosshair through the origin.
}
