% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_col_default.R
\name{gg_col_default}
\alias{gg_col_default}
\title{Select ggplot default colour palette of n. colours}
\usage{
gg_col_default(nColours = 3, plot = TRUE)
}
\arguments{
\item{nColours}{N colours to display}

\item{plot}{TRUE: plot the colour palette; or FALSE: only returns the hex codes in the console.}
}
\description{
Select ggplot default colour palette of n. colours
}
