% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_big.R
\name{head_big}
\alias{head_big}
\title{big head}
\usage{
head_big(.df, n = 6)
}
\arguments{
\item{.df}{A data frame, tibble or matrix}

\item{n}{number of rows and columns to print (square output, starting from row1, col1)}
}
\value{
A data_frame (aka tibble) to the console. EG.GO and entrez_joinTable are exported to R_GlobalEnv.
}
\description{
Extract from a DGE object genes associated with a GO term of interest, e.g. from GSEA testing. Requires GOsource() output.
}
\examples{
mtc_mat <- as.matrix(mtcars)
mtc_mat |> head_big(8)
}
