% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_bottom.R
\name{legend_bottom}
\alias{legend_bottom}
\alias{bottom_legend}
\title{Legend bottom}
\usage{
legend_bottom(ncol = NULL, invert_rows = FALSE)
}
\arguments{
\item{ncol}{Integer; number of columns in the legend (default: NULL, no specific column arrangement).}

\item{invert_rows}{Logical; if TRUE, reverses the row order of legend items (default: FALSE).}
}
\value{
A list of ggplot2 theme and guide specifications.
}
\description{
Adjusts the ggplot2 legend position to the bottom of the plotting space.
Optionally, you can specify the number of columns in the legend and whether to invert the row order of legend items.
For more details, see \href{https://github.com/bansell/tidyExt#tidyext-package}{tidyExt vignette}.
}
\examples{
if(require("ggplot2")) {
  ggplot(mpg, aes(x = cty, y = hwy)) +
    geom_point(aes(col = manufacturer)) +
    legend_bottom()

  ggplot(mpg, aes(x = cty, y = hwy)) +
    geom_point(aes(col = manufacturer)) +
    legend_bottom(ncol = 2, invert_rows = TRUE)
}
}
