% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_hide.R
\name{legend_hide}
\alias{legend_hide}
\title{No legend}
\usage{
legend_hide()
}
\description{
Omits ggplot2 legend.
}
\details{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\examples{
require(ggplot2)
mpg |> ggplot2::ggplot(aes(x=cty,y=hwy)) + geom_point(aes(col=drv)) + legend_hide()
}
