% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_prop.R
\name{sort_prop}
\alias{sort_prop}
\title{Count any number of columns, print percentage and sort descending.}
\usage{
sort_prop(tbl, ...)
}
\arguments{
\item{tbl}{A data frame/tibble}

\item{...}{Any number of columns to count}
}
\description{
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
