% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlab_angle.R
\name{xlab_angle}
\alias{xlab_angle}
\alias{x_angle}
\title{Set x axis labels at any angle}
\usage{
xlab_angle(x = 30)
}
\arguments{
\item{x}{angle in degrees}
}
\description{
Set x axis labels at any angle
}
\examples{
require(ggplot2)
mpg |> ggplot2::ggplot(aes(x=manufacturer,y=hwy)) + geom_boxplot() + xlab_angle(45)
}
