% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ylab_angle.R
\name{y_angle}
\alias{y_angle}
\title{Set y axis labels at any angle}
\usage{
y_angle(y = 30)
}
\arguments{
\item{y}{angle in degrees}
}
\description{
Set y axis labels at any angle
}
\examples{
require(ggplot2)
mpg |> ggplot2::ggplot(aes(x=hwy,y=manufacturer)) + geom_boxplot() + y_angle(45)
}
