% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_helpers.R
\name{all_after}
\alias{all_after}
\alias{all_before}
\title{stringr regex helpers}
\usage{
all_after(character)

all_before(character)
}
\arguments{
\item{character}{A character string.}
}
\value{
A vector.
}
\description{
Identifies sub-string occuring before or after character.
Note: these are technically all_after_last() and all_before_first() occurrences.
see \href{https://github.com/bansell/tidyExt#tidyext-package}{\strong{tidyExt vignette}}
}
\section{Functions}{
\itemize{
\item \code{all_after()}: Extract all characters after match.

\item \code{all_before()}: Extract all characters before match.

}}
