% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as_character}
\alias{as_character}
\alias{as_numeric}
\alias{as_integer}
\alias{as_factor}
\alias{as_logical}
\alias{as_complex}
\title{as_character}
\usage{
as_character(x)

as_numeric(x)

as_integer(x)

as_factor(x)

as_logical(x)

as_complex(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of the specified data type.
}
\description{
Data type conversion wrappers
}
\details{
Simple wrapper for as.character() for consistency with tidy syntax.
Note: these are technically all_after_last() and all_before_first() occurrences.
see \href{https://vctrs.r-lib.org/reference/vec_cast.html}{\strong{vctrs::vec_cast()}} for a complete treatment of type conversion/coersion.
}
\section{Functions}{
\itemize{
\item \code{as_character()}: Convert vector to character data type

\item \code{as_numeric()}: Convert vector to numeric data type

\item \code{as_integer()}: Convert vector to integer data type

\item \code{as_factor()}: Convert vector to factor data type

\item \code{as_logical()}: Convert vector to logical data type

\item \code{as_complex()}: Convert vector to complex data type

}}
\examples{
vec_numeric <- c(3.142,2.5,3.4)
vec_character <- c('mitsubishi','subaru','honda')
vec_integer <- c(1,0,1)
vec_logical <- c(TRUE,FALSE,TRUE)

as_character(vec_numeric)
as_integer(vec_numeric)
as_factor(vec_character)
as_logical(vec_integer)
as_numeric(vec_logical)
}
